+++ src/object/getPropByList.ts
@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
+
 /**
  * 根据属性列表获取对象属性
  * @param {object} value 对应对象
@@ -6,13 +6,13 @@
  * @param {boolean} [showError] 显示错误输出
  * @returns
  */
-function getPropByList(value: Record<PropertyKey, any>, propList: string[] | number[], showError?: boolean): any {
+function getPropByList(value: Record<PropertyKey, unknown>, propList: string[] | number[], showError?: boolean): unknown {
   let data = value
   try {
     for (let n = 0; n < propList.length; n++) {
       const prop = propList[n]
       if (prop || prop === 0) {
-        data = data[prop]
+        data = data[prop] as Record<PropertyKey, unknown>
       }
     }
     return data